+++ doc/version-history.rst
@@ -8,6 +8,7 @@ v0.9.0
 ------
 
 * Support for automatic FCU's heaters PWM (~power) control.
+* VMSLogger Kafka
 
 v0.8.3
 ------